pyproject.toml
[project]
name = "hexapod"
version = "0.1.0"
description = "Hexapod robot control software"
requires-python = ">=3.11"
dependencies = [
    "pyserial>=3.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hexapod"]
